+++ golang/testsuite/testsuite_impl/exec_command_test/exec_command_test_.go
@@ -1,6 +1,7 @@
 package exec_command_test
 
 import (
+	"fmt"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/testsuite"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/services_impl/exec_cmd_test"
@@ -10,8 +11,10 @@ import (
 )
 
 const (
-	execCmdTestImage = "alpine:3.12.4"
-	testServiceId = "test"
+	execCmdTestImage      = "alpine:3.12.4"
+	inputForLogOutputTest = "hello"
+	expectedLogOutput = "hello\n"
+	testServiceId         = "test"
 
 	successExitCode int32 = 0
 
@@ -21,6 +24,11 @@ const (
 
 var execCommandThatShouldWork = []string{
 	"true",
+}
+
+var execCommandThatShouldHaveLogOutput = []string{
+	"echo",
+	inputForLogOutputTest,
 }
 
 var execCommandThatShouldFail = []string{
@@ -61,7 +69,7 @@ func (e ExecCommandTest) Run(uncastedNetwork networks.Network, testCtx testsuite
 	castedService := uncastedService.(*exec_cmd_test.ExecCmdTestService)
 
 	logrus.Infof("Running exec command '%v' that should return a successful exit code...", execCommandThatShouldWork)
-	shouldWorkExitCode, err := castedService.RunExecCmd(execCommandThatShouldWork)
+	shouldWorkExitCode, _, err := castedService.RunExecCmd(execCommandThatShouldWork)
 	if err != nil {
 		testCtx.Fatal(stacktrace.Propagate(err, "An error occurred running exec command '%v'", execCommandThatShouldWork))
 	}
@@ -72,13 +80,27 @@ func (e ExecCommandTest) Run(uncastedNetwork networks.Network, testCtx testsuite
 
 
 	logrus.Infof("Running exec command '%v' that should return an error exit code...", execCommandThatShouldFail)
-	shouldFailExitCode, err := castedService.RunExecCmd(execCommandThatShouldFail)
+	shouldFailExitCode, _, err := castedService.RunExecCmd(execCommandThatShouldFail)
 	if err != nil {
 		testCtx.Fatal(stacktrace.Propagate(err, "An error occurred running exec command '%v'", execCommandThatShouldFail))
 	}
 	if shouldFailExitCode == successExitCode {
-		testCtx.Fatal(stacktrace.NewError("Exec command '%v' should fail, but got successful exit code %v", shouldFailExitCode, successExitCode))
+		testCtx.Fatal(stacktrace.NewError("Exec command '%v' should fail, but got successful exit code %v", execCommandThatShouldFail, successExitCode))
 	}
+
+	logrus.Infof("Running exec command '%v' that should return log output...", execCommandThatShouldHaveLogOutput)
+	shouldHaveLogOutputExitCode, logOutput, err := castedService.RunExecCmd(execCommandThatShouldHaveLogOutput)
+	if err != nil {
+		testCtx.Fatal(stacktrace.Propagate(err, "An error occurred running exec command '%v'", execCommandThatShouldHaveLogOutput))
+	}
+	if shouldHaveLogOutputExitCode != successExitCode {
+		testCtx.Fatal(stacktrace.NewError("Exec command '%v' should work, but got unsuccessful exit code %v", execCommandThatShouldHaveLogOutput, shouldHaveLogOutputExitCode))
+	}
+	logOutputStr := fmt.Sprintf("%s", *logOutput)
+	if logOutputStr != expectedLogOutput {
+		testCtx.Fatal(stacktrace.NewError("Exec command '%v' should return %v, but got %v.", execCommandThatShouldHaveLogOutput, inputForLogOutputTest, logOutputStr))
+	}
+
 	logrus.Info("Exec command returned error exit code as expected")
 }